requirements.txt
transformers
torch
pillow
tqdm
pandas
datasets
peft
accelerate

